Collage helper done
Collage helper is a module responsable to generate the collage of images.
It expects a hash with some parameters (it will be explained on README file).

Collage module relies on RMagick, which relies on ImageMagick.
So, it's mandatory the instalation of imagemagick on host.

Also, the method used to generate the collage requires a Freetype font, which
isn't installed by default on some operating systems.

So, a script were created to download ghostscript font, unpack it, install and
add reference to ImageMagick (Linux only). The script is saved in bin/install-imagemagick-font.

Instruction on how to execute the script will be described on README file as well.